Update plotting functions in graphstates.py
Change the plotting methods draw (in EGrap) and sketch (in CVGraph). First, obtain the font style from the new self.font_props in the EGraph.  Small modification also to exceptions and error messages. Get sketch to output a matplotlib axes object for the benefit of e.g. the decoding functions.