Add MWPM function to decoder.py
Add a function for minimum weight perfect matching to dedocder.py. This runs max_weight_matching on an a matching graph, looking in the inverse_weight attribute for the weights. User can also draw the matching on the syndrome plot produced in the deocding graph. (Note for now this requires decoding_graph to run the plotting functions followed immediately by MWPM - this will be fixed later).