Add CV_decoder to decoder.py and remove translator from CVGraph class
Since CV-to-DV translation is a feature of the decoder and not the CVGraph per se, remove the translator attribute and translate_outcomes method from CVGraph, and create a CV_decoder function in decoder.py. On the other hand, keep bit_values property in CVGraph so that the bit values can be plotted if they have already been translated.

Additionally, minor changes to some docstrings.