Update monte_carlo in simulations.py
Update monte_carlo function to take advantage of the new 'correct' function from decoder.py, and to output the failure probability and the number-of-trial-dependent error. For now, the actual decoding steps are commented out (until the recovery and check_correct functions are filled in). For preliminary testing, the success of a trial is a random binary digit.