Add non-periodic boundary condition options to RHG_graph
Include the possibility of different non-periodic boundary conditions to RHG_Graph. For now, only the right, top, and back boundaries are changed; the left, bottom, and front boundaries are 'primal.'  If boundaries is 'primal', all primal (the default). If 'dual', all dual. If None, everything primal except for the back slice, which is dual.