Improve construction of RHG_lattice in RHG.py
Change the way the RHG lattice is constructed. Before, it was done in terms of data and ancillae qubits. Now, it is done in terms of primal and dual qubits (all the primal qubits are identified, and then their dual neighbours are populated). This way is faster and less redundant, and gives the graph useful attributes for later use. Other changes:

- EGraph is now informed that their are macronodes through the 'macronodes' rather than the 'indexer' argument.
- Polarity is no longer a boolean but a function that specifies the polarity (e.g. the newly added alternating_polarity, which gives a special +1/-1 alternating edge weight pattern)
- If memory_saver is False, each node gets a "type" and "color" attribute. Type can be primal or dual, and color is determined by the dictionary in the node_color argument.
- If memory_saver is False, each edge gets a "color" attribute, determined by the dictionary in the edge_color argument
- A list of the coordinates of 6-body stabilizers is stored in the "primal_cubes" graph attribute of the EGraph.