Update README.md
Visual Studio compilers and prompts were causing a variety of errors for different users and needed pybind, compilers, and system path settings -- an overkill for sure with large files to be installed. 

As discussed with Max, I have now found a robust and quick way to install using Windows through WSL (no cheating though), which we can confidently support for all users.