Fix sometimes failing test_hybridize
We are basically checking here whether a sampling from a binomial distribution for populating the lattice with p-squeezed states results in a number of p-squeezed states close to p_swap * (size of lattice). However, for one of the tests, p_swap is taken randomly between 0 and 1 - if it is too close to 0 (as it is here), the the isclose function is too unforgiving. Rather than playing around with rtol and atol, I propose to shift the random sample so that the minimal p_swap is 0.01.