Allow instantiation of CVLayer with code object + misc. changes (#25)
* Initialize CVLayer with code

Allow for user to supply an EGraph or a code object to CVLayer. Makes it easier for user.

* Update CVLayer arguments

(In light of previous change)

* Fix sometimes failing test_hybridize

We are basically checking here whether a sampling from a binomial distribution for populating the lattice with p-squeezed states results in a number of p-squeezed states close to p_swap * (size of lattice). However, for one of the tests, p_swap is taken randomly between 0 and 1 - if it is too close to 0 (as it is here), the the isclose function is too unforgiving. Rather than playing around with rtol and atol, I propose to shift the random sample so that the minimal p_swap is 0.01.

* Update CHANGELOG.md

* Update pull_request_template.md

* Update CHANGELOG.md

* Add new test to test_graphstates

Simple test to check initialization of CVLayer with a SurfaceCode object

* Create codecov.yml

* Update CHANGELOG.md

* Update _version.py

Co-authored-by: Narimanium <25132802+nariman87@users.noreply.github.com>